Refactor CSS and update components and structure
