+++ journals/2024_05_02.md
@@ -1,1 +1,5 @@
+- - [ ] Figma Page 1 #37
+	- - See commit: [bb9f292](https://github.com/infinitepixel-dev/InfinitePixel-Main/commit/bb9f2923234ab2a504f5ac008d32f11319c00ed3)
+	- Updated .gitignore to exclude logseq directory
+	- {{link-tasks}}
 -
