+++ src/components/pages/HeroPage/heroBackground.jsx
