+++ src/components/pages/SplashPage/splashPage.css
