+++ src/components/pages/SplashPage/splashPage.jsx
